refactor(new reviewer): move key processing to viewModel
